+++ permission_handler_apple/ios/Classes/strategies/AssistantPermissionStrategy.m
@@ -12,8 +12,7 @@
 @implementation AssistantPermissionStrategy
 
 - (PermissionStatus)checkPermissionStatus:(PermissionGroup)permission {
-    INSiriAuthorizationStatus assistantPermission = [INPreferences siriAuthorizationStatus];
-    return [AssistantPermissionStrategy parsePermission:assistantPermission];
+    return PermissionStatusDenied;
 }
 
 - (void)checkServiceStatus:(PermissionGroup)permission completionHandler:(ServiceStatusHandler)completionHandler {
@@ -21,30 +20,9 @@
 }
 
 - (void)requestPermission:(PermissionGroup)permission completionHandler:(PermissionStatusHandler)completionHandler errorHandler:(PermissionErrorHandler)errorHandler {
-    PermissionStatus status = [self checkPermissionStatus:permission];
-    if (status != PermissionStatusDenied) {
-        completionHandler(status);
-        return;
-    }
-
-    [INPreferences requestSiriAuthorization:^(INSiriAuthorizationStatus status) {
-        PermissionStatus permissionStatus = [AssistantPermissionStrategy parsePermission:status];
-        completionHandler(permissionStatus);
-    }];
+    completionHandler(PermissionStatusDenied);
 }
 
-+ (PermissionStatus)parsePermission:(INSiriAuthorizationStatus)assistantPermission API_AVAILABLE(ios(10)){
-    switch(assistantPermission){
-        case INSiriAuthorizationStatusAuthorized:
-            return PermissionStatusGranted;
-        case INSiriAuthorizationStatusRestricted:
-            return PermissionStatusRestricted;
-        case INSiriAuthorizationStatusDenied:
-            return PermissionStatusPermanentlyDenied;
-        case INSiriAuthorizationStatusNotDetermined:
-            return PermissionStatusDenied;
-    }
-}
 @end
 
 #else