+++ permission_handler/CHANGELOG.md
@@ -32,7 +32,7 @@
 
 ## 11.2.0
 
-- Adds the `Permission.assistant` which allows users to request permissions to access SiriKit on iOS and macOS platforms. This is a no-op on all other platforms.
+- Adds the `Permission.assistant` permission on iOS and macOS platforms. This is a no-op on all other platforms.
 
 ## 11.1.0