+++ permission_handler_html/CHANGELOG.md
@@ -28,7 +28,7 @@
 
 ## 0.1.1
 
-- Updates the dependency on `permission_handler_platform_interface` to version 4.1.0 (SiriKit support is only available for iOS and macOS).
+- Updates the dependency on `permission_handler_platform_interface` to version 4.1.0.
 
 ## 0.1.0+2